Save one working version.
